fix: move command line specific options to cli.go, keep appimagetool's logic in appimagetool.go
